Use FunctionType instead of FunctionDeclaration for createLLVMIntrinsic
