+++ projects/com.oracle.truffle.llvm.parser.base/src/com/oracle/truffle/llvm/parser/base/model/LLVMToBitcodeAdapter.java
@@ -324,7 +324,7 @@ public final class LLVMToBitcodeAdapter {
         return funcDef;
     }
 
-    public static FunctionDeclaration resolveFunctionHeader(LLVMParserRuntime runtime, FunctionHeader header) {
+    public static FunctionType resolveFunctionHeader(LLVMParserRuntime runtime, FunctionHeader header) {
         Type returnType = resolveType(runtime.resolve(header.getRettype()));
         List<Type> args = new ArrayList<>();
         boolean hasVararg = false;
@@ -337,9 +337,8 @@ public final class LLVMToBitcodeAdapter {
             }
         }
         FunctionType funcType = new FunctionType(returnType, args.toArray(new Type[args.size()]), hasVararg);
-        FunctionDeclaration funcDecl = new FunctionDeclaration(funcType);
-        funcDecl.setName(header.getName().substring(1));
-        return funcDecl;
+        funcType.setName(header.getName().substring(1));
+        return funcType;
     }
 
     private static Linkage resolveLinkage(String linkage) {
