+++ projects/com.oracle.truffle.llvm.parser.base/src/com/oracle/truffle/llvm/parser/base/model/functions/FunctionDeclaration.java
@@ -32,30 +32,13 @@ package com.oracle.truffle.llvm.parser.base.model.functions;
 import com.oracle.truffle.llvm.parser.base.model.symbols.ValueSymbol;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.Constant;
 import com.oracle.truffle.llvm.parser.base.model.types.FunctionType;
-import com.oracle.truffle.llvm.parser.base.model.types.PointerType;
-import com.oracle.truffle.llvm.parser.base.model.types.Type;
 
-public final class FunctionDeclaration extends FunctionType implements Constant, ValueSymbol {
+public final class FunctionDeclaration extends FunctionType implements Constant {
 
     private String name = ValueSymbol.UNKNOWN;
 
     public FunctionDeclaration(FunctionType type) {
         super(type.getReturnType(), type.getArgumentTypes(), type.isVarArg());
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public Type getType() {
-        return new PointerType(super.getType());
-    }
-
-    @Override
-    public void setName(String name) {
-        this.name = "@" + name;
     }
 
     @Override
