+++ projects/com.oracle.truffle.llvm.parser.base/src/com/oracle/truffle/llvm/parser/base/model/types/FunctionType.java
@@ -31,15 +31,18 @@ package com.oracle.truffle.llvm.parser.base.model.types;
 
 import com.oracle.truffle.llvm.parser.LLVMBaseType;
 import com.oracle.truffle.llvm.parser.base.datalayout.DataLayoutConverter;
+import com.oracle.truffle.llvm.parser.base.model.symbols.ValueSymbol;
 import com.oracle.truffle.llvm.types.LLVMFunctionDescriptor;
 
-public class FunctionType implements Type {
+public class FunctionType implements Type, ValueSymbol {
 
     private final Type type;
 
     private final Type[] args;
 
     private final boolean isVarArg;
+
+    private String name;
 
     public FunctionType(Type type, Type[] args, boolean isVarArg) {
         this.type = type;
@@ -54,6 +57,11 @@ public class FunctionType implements Type {
     @Override
     public LLVMBaseType getLLVMBaseType() {
         return LLVMBaseType.FUNCTION_ADDRESS;
+    }
+
+    @Override
+    public Type getType() {
+        return new PointerType(Type.super.getType());
     }
 
     public Type getReturnType() {
@@ -81,6 +89,16 @@ public class FunctionType implements Type {
     @Override
     public int getBits() {
         return 0;
+    }
+
+    @Override
+    public void setName(String name) {
+        this.name = "@" + name;
+    }
+
+    @Override
+    public String getName() {
+        return name;
     }
 
     @Override
